refactor: move PointCloudClassification to autoware_point_types
